+++ src/main/java/com/animania/common/entities/rodents/ai/EntityAIFindFoodRabbits.java
@@ -5,6 +5,7 @@ import java.util.Random;
 import com.animania.common.entities.rodents.rabbits.EntityAnimaniaRabbit;
 import com.animania.common.handler.BlockHandler;
 import com.animania.common.tileentities.TileEntityTrough;
+import com.animania.config.AnimaniaConfig;
 
 import net.minecraft.block.Block;
 import net.minecraft.entity.EntityCreature;
@@ -112,7 +113,9 @@ public class EntityAIFindFoodRabbits extends EntityAIBase
 					ech.setFed(true);
 				} 
 				
-				temptedEntity.world.destroyBlock(currentpos, false);
+				if (AnimaniaConfig.gameRules.plantsRemovedAfterEating) {
+					temptedEntity.world.destroyBlock(currentpos, false);
+				}
 				
 				return false;
 			}
