+++ src/main/java/com/animania/common/entities/sheep/ai/EntityAIFindFoodSheep.java
@@ -5,6 +5,7 @@ import java.util.Random;
 import com.animania.common.entities.sheep.EntityAnimaniaSheep;
 import com.animania.common.handler.BlockHandler;
 import com.animania.common.tileentities.TileEntityTrough;
+import com.animania.config.AnimaniaConfig;
 
 import net.minecraft.block.Block;
 import net.minecraft.entity.EntityCreature;
@@ -111,7 +112,9 @@ public class EntityAIFindFoodSheep extends EntityAIBase
 					ech.setFed(true);
 				} 
 				
-				temptedEntity.world.destroyBlock(currentpos, false);
+				if (AnimaniaConfig.gameRules.plantsRemovedAfterEating) {
+					temptedEntity.world.destroyBlock(currentpos, false);
+				}
 				
 				return false;
 			}
